requirements.txt
Flask
Werkzeug
Pillow
